修改订单api，分页，按最近时间排序； 修改权限菜单api，保证菜单的删除添加，同步关联角色； 商品只需要二级分类，三级分类可选； 更新admin管理后台；

+++ routes/admin/menu.js
@@ -10,25 +10,36 @@ let db = require('../../config/mysql');
  * @apiPermission admin
  *
  * @apiParam {String} name 分类名称.
+ * @apiParam {String} component 关联组件名称.
  * @apiParam {Number} pId 父级id.
  * @apiParam {String} path 菜单url地址.
  * @apiParam {String} order 菜单显示顺序，按照数字从小到大排序，如2001.
- * @apiParam {String} component 组件名称.
  *
  * @apiSampleRequest /api/menu
  */
 router.post("/", function(req, res) {
-	let { name, pId, path, order, component } = req.body;
-	let sql = `INSERT INTO MENU (name,pId,path,menu_order,component) VALUES (?,?,?,?,?) `;
-	db.query(sql, [name, pId, path, order, component], function(results, fields) {
-		//成功
-		res.json({
-			status: true,
-			msg: "success!",
-			data: {
-				id: results.insertId
+	let { name, pId, component, path, order } = req.body;
+	let sql = `INSERT INTO MENU (name,pId,component,path,menu_order) VALUES (?,?,?,?,?) `;
+	db.query(sql, [name, pId, component, path, order], function({ insertId }, fields) {
+		let sql = `INSERT INTO role_menu (role_id,menu_id) VALUES (1,?)`;
+		db.query(sql, [insertId], function(results, fields) {
+			if (!results.affectedRows) {
+				res.json({
+					status: false,
+					msg: "创建失败！"
+				});
+				return;
 			}
-		});
+			//成功
+			res.json({
+				status: true,
+				msg: "创建成功!",
+				data: {
+					id: insertId
+				}
+			});
+		})
+
 	});
 });
 /**
@@ -43,12 +54,19 @@ router.post("/", function(req, res) {
  */
 router.delete("/", function(req, res) {
 	let { id } = req.query;
-	let sql = `DELETE FROM MENU WHERE id = ?`;
-	db.query(sql, [id], function(results, fields) {
+	let sql = `DELETE FROM MENU WHERE id = ?;DELETE FROM ROLE_MENU WHERE menu_id = ?`;
+	db.query(sql, [id, id], function(results, fields) {
+		if (!results[1].affectedRows) {
+			res.json({
+				status: false,
+				msg: "删除失败！"
+			});
+			return;
+		}
 		//成功
 		res.json({
 			status: true,
-			msg: "success!"
+			msg: "删除成功!"
 		});
 	});
 });
@@ -58,17 +76,26 @@ router.delete("/", function(req, res) {
  * @apiGroup admin-Menu
  * @apiPermission admin
  *
- * @apiParam { Number } id 子菜单id.
- * @apiParam { String } name 子菜单名称.
- * @apiParam { String } path 子菜单url地址.
- * @apiParam {String} component 组件名称.
+ * @apiParam {Number} id 子菜单id.
+ * @apiParam {String} name 分类名称.
+ * @apiParam {String} component 关联组件名称.
+ * @apiParam {Number} pId 父级id.
+ * @apiParam {String} path 菜单url地址.
+ * @apiParam {String} order 菜单显示顺序，按照数字从小到大排序，如2001.
  *
  * @apiSampleRequest /api/menu
  */
 router.put("/", function(req, res) {
-	let { name, path, id, order, component } = req.body;
-	let sql = `UPDATE MENU SET name = ? , path = ?, menu_order = ?, component = ? WHERE id = ? `;
-	db.query(sql, [name, path, order, component, id], function(results, fields) {
+	let { id, name, pId, component, path, order } = req.body;
+	let sql = `UPDATE MENU SET name = ?,pId = ?,component = ?, path = ?, menu_order = ? WHERE id = ? `;
+	db.query(sql, [name, pId, component, path, order, id], function(results, fields) {
+		if (!results.affectedRows) {
+			res.json({
+				status: false,
+				msg: "failed！"
+			});
+			return;
+		}
 		//成功
 		res.json({
 			status: true,
@@ -91,6 +118,13 @@ router.put("/icon", function(req, res) {
 	let { id, icon } = req.body;
 	let sql = `UPDATE MENU SET icon_id = ? WHERE id = ? `;
 	db.query(sql, [icon, id], function(results, fields) {
+		if (!results.affectedRows) {
+			res.json({
+				status: false,
+				msg: "failed！"
+			});
+			return;
+		}
 		//成功
 		res.json({
 			status: true,
@@ -110,7 +144,7 @@ router.put("/icon", function(req, res) {
  */
 router.get("/sub", function(req, res) {
 	let sql =
-		`SELECT m.id,m.name,m.pId,m.path,m.component, m.menu_order AS 'order', i.name AS 'icon' FROM MENU m LEFT JOIN icon i ON m.icon_id = i.id WHERE m.pId = ? ORDER BY m.menu_order`;
+		`SELECT m.id, m.name, m.pId, m.path, m.component, m.menu_order AS 'order', i.name AS 'icon' FROM MENU m LEFT JOIN ICON i ON m.icon_id = i.id WHERE m.pId = ? ORDER BY m.menu_order`;
 	db.query(sql, [req.query.pId], function(results, fields) {
 		//成功
 		res.json({
@@ -133,12 +167,12 @@ router.get("/sub", function(req, res) {
 router.get('/tree', function(req, res) {
 	let { id } = req.query;
 	let sql =
-		`SELECT m.*,i.name AS 'icon' FROM MENU m JOIN role_menu rm ON rm.menu_id = m.id LEFT JOIN icon i ON m.icon_id = i.id WHERE rm.role_id = ? ORDER BY menu_order;`;
+		`SELECT m.*,i.name AS 'icon' FROM MENU m JOIN role_menu rm ON rm.menu_id = m.id LEFT JOIN ICON i ON m.icon_id = i.id WHERE rm.role_id = ? ORDER BY m.menu_order;`;
 	db.query(sql, [id], function(results) {
 		//筛选出一级菜单
-		let menu_1st = results.filter((item) => item.pId === 1 ? item : null);
+		let cate_1st = results.filter((item) => item.pId === 1 ? item : null);
 		//递归循环数据
-		parseToTree(menu_1st);
+		parseToTree(cate_1st);
 		//递归函数
 		function parseToTree(array) {
 			array.forEach(function(parent) {
@@ -155,7 +189,7 @@ router.get('/tree', function(req, res) {
 		res.json({
 			status: true,
 			msg: "success!",
-			data: menu_1st
+			data: cate_1st
 		});
 	});
 });
